ch25) Clear posts in user when posts are deleted

+++ ch24/controllers/feed.js
@@ -3,6 +3,7 @@ import path from 'path'
 import { validationResult } from 'express-validator'
 
 const Post = require('../models/post')
+const User = require('../models/user')
 
 exports.getPosts = (req, res, next) => {
   const currentPage = req.query.page || 1
@@ -50,13 +51,23 @@ exports.createPost = (req, res, next) => {
     title: title,
     content: content,
     imageUrl: imageUrl,
-    creator: { name: 'novdov' }
+    creator: req.userId
   })
+  let creator
   post.save()
+    .then(result => {
+      return User.findById(req.userId)
+    })
+    .then(user => {
+      creator = user
+      user.posts.push(post)
+      return user.save()
+    })
     .then(result => {
       res.status(201).json({
         message: 'Post created successfully',
-        post: result
+        post: post,
+        creator: { _id: creator._id, name: creator.name }
       })
     })
     .catch(err => {
@@ -115,6 +126,9 @@ exports.updatePost = (req, res, next) => {
         error.statusCode = 404
         throw error
       }
+      if (post.creator.toString() !== req.userId) {
+        throw new Error('Not authenticated')
+      }
       if (imageUrl !== post.image) {
         clearImage(post.imageUrl)
       }
@@ -144,12 +158,21 @@ exports.deletePost = (req, res, next) => {
         error.statusCode = 404
         throw error
       }
+      if (post.creator.toString() !== req.userId) {
+        throw new Error('Not authenticated')
+      }
       // Check logged in user
       clearImage(post.imageUrl)
       return Post.findByIdAndRemove(postId)
     })
     .then(result => {
-      console.log(result)
+      return User.findById(req.userId)
+    })
+    .then(user => {
+      user.posts.pull(postId)
+      return user.save()
+    })
+    .then(result => {
       res.status(200).json({ message: 'Deleted post' })
     })
     .catch(err => {
